логика транзакции| распределили классы для баз данных по каталогам
