+++ src/main/java/hitbtc/InfoDB/Coins/Coins.java
@@ -1,4 +1,4 @@
-package hitbtc.InfoDB;
+package hitbtc.InfoDB.Coins;
 
 import DB.DBconnactionVPS;
 import DB.GetVal;
@@ -47,23 +47,3 @@ public class Coins {
     }
 }
 
-class Coin {
-    public int id;
-    public String abbreviation;
-    public String name;
-    public String description;
-
-    Coin(int id, String abbreviation, String name, String description){
-        this.id = id;
-        this.abbreviation = abbreviation;
-        this.name = name;
-        this.description = description;
-    }
-
-    public String toString(){
-        String result = new StringBuilder().append("id:").append(id).append("\tabbreviation:").append(abbreviation)
-                .append("\tname:").append(name).append("\tdescription:").append(description).toString();
-        return result;
-    }
-
-}
